Move some utils to internal header
